requirements.txt
customtkinter
bcrypt
authlib[httpx]
python-dotenv
Pillow
moviepy
pydub
PyJWT>=2.8.0
cryptography>=41.0.0
keyring>=24.0.0
email-validator>=2.0.0
pyotp>=2.9.0
ratelimit>=2.2.1